optional-requirements.txt
celery>=3.1,<4.0
boto>=2.46
django-celery>=3.1.0
xlsxwriter==0.8.5
factory-boy==2.8.1
mock
